Tests for Virtual Memory
Memory implements three ranges, more to be added.

Uses a hash table to simulate memory. Uses a stack to keep track of freed memory, and uses stack if possible, if not, keeps incrementing addr pointer until it reaches the limit.

If all memory is "freed", we assume it is free and overwrite it by resetting the pointer